🍗 :: [#349] pose_main_screen 리펙토링
Class명 변경과 변경된 Class명 변경
